\name{CVEP_JM}
\alias{CVEP_JM}
\title{Joint Modelling Procedure in Analyzing Highly Unbalanced CVEP Data}
\usage{
CVEP_JM()
}
\description{
Joint modelling (JM) procedure which analyze highly unbalanced data from CVEP, in which some standard varieties, called controls, are assigned by the program.
}
\examples{
CVEP_JM()
}
